linux: don't require libnuma anymore
Many users reported failures to get mbind support because libnuma devel
headers are not installed. We only need libnuma because that's where
numa syscalls are wired, let's hardwire them to remove the dependency.

We still use libnuma for the specific linux-libnuma.h testing,
but not in the core anymore.

The main drawback is that we now always use the syscall() macro
(previously only used when libc didn't define sched_setaffinity),
that means we require a working syscall with 6 arguments (for mbind),
which works since at least glibc 2.11 (2009/11) according to libnuma.
At least it seems to work in Debian 6, Ubuntu 8.04 and CentOS 5.9.

If we really need support for older libc, we may have to hardwire
syscall6().

Another drawback is that we assume all these functions and flags
are available. Those were added to the kernel in 2006, so it's fine.
If using an older kernel, support will say that things are supported
but we may actually return ENOSYS.

By the way, fix the way we check syscall() in configure
(syscallX needs X+1 arguments).